Fix instruction cache invalidation in sys mprotect
Resolve host pointers with host_ptr_for_gpa for high VA and
g->host_base + mprot_off for primary memory in sys_mprotect.

Run instruction cache invalidation on the mprotect fast path
when LINUX_PROT_EXEC is set.

Add test-jit-icache to test-matrix.sh and manifest.txt.

+++ src/syscall/internal.h
@@ -952,7 +952,7 @@ static inline fd_block_state_t fd_block_state_of(const fd_entry_t *e)
     return (fd_block_state_t) {
         .type = e->type,
         .generation = e->generation,
-        .seals = e->seals,
+        .seals = (unsigned int) e->seals,
         .can_block = e->can_block,
         .nonblock_owned = e->nonblock_owned,
         .guest_nonblock = (e->linux_flags & LINUX_O_NONBLOCK) != 0,
